CI: Overhaul Windows setup script
Previously, there was no way of forcing a re-download
of cached dependencies when they were replaced by new ones. In
addition, there was really no verification of downloads or cache.
Now, changing a few lines at the top of the file will automagically
force a cache update.

+++ .travis/setup-windows.sh
@@ -1,11 +1,5 @@
 #!/bin/sh -ex
 
-# Git bash doesn't have rev, so here it is
-rev()
-{
-    echo "$1" | awk '{ for(i = length($0); i != 0; --i) { a = a substr($0, i, 1); } } END { print a }'
-}
-
 # These are Azure specific, so we wrap them for portability
 REPO_NAME="$BUILD_REPOSITORY_NAME"
 REPO_BRANCH="$SYSTEM_PULLREQUEST_SOURCEBRANCH"
@@ -22,56 +16,82 @@ QT_BASE_URL="${QT_HOST}${QT_PREFIX}qtbase${QT_SUFFIX}"
 QT_WINE_URL="${QT_HOST}${QT_PREFIX}qtwinextras${QT_SUFFIX}"
 QT_DECL_URL="${QT_HOST}${QT_PREFIX}qtdeclarative${QT_SUFFIX}"
 QT_TOOL_URL="${QT_HOST}${QT_PREFIX}qttools${QT_SUFFIX}"
-LLVMLIBS='https://github.com/RPCS3/llvm-mirror/releases/download/custom-build-win/llvmlibs_mt.7z'
-GLSLANG='https://github.com/RPCS3/glslang/releases/download/custom-build-win/glslanglibs_mt.7z'
+LLVMLIBS_URL='https://github.com/RPCS3/llvm-mirror/releases/download/custom-build-win/llvmlibs_mt.7z'
+LLVM_SHA="773273312b6bfe0c361b3780e175ab2935f8d10558431420dee2a1414ee965d0"
+GLSLANG_URL='https://github.com/RPCS3/glslang/releases/download/custom-build-win/glslanglibs_mt.7z'
+GLSLANG_SHA="b2f47a20239eddd423257c10258e1c7889a5bb7d66773460e3a87b4dbead7c1d"
 VULKAN_SDK_URL="https://sdk.lunarg.com/sdk/download/${VULKAN_VER}/windows/VulkanSDK-${VULKAN_VER}-Installer.exe"
 
-QT_URLS="          \
+DEP_URLS="         \
     $QT_BASE_URL   \
     $QT_WINE_URL   \
     $QT_DECL_URL   \
-    $QT_TOOL_URL"
+    $QT_TOOL_URL   \
+    $LLVMLIBS_URL  \
+    $GLSLANG_URL   \
+    $VULKAN_SDK_URL"
 
 # Azure pipelines doesn't make a cache dir if it doesn't exist, so we do it manually
 [ -d "$CACHE_DIR" ] || mkdir "$CACHE_DIR"
 
 # Pull all the submodules except llvm, since it is built separately and we just download that build
-# Do not quote the subshell
 # Note: Tried to use git submodule status, but it takes over 20 seconds
+# shellcheck disable=SC2046
 git submodule -q update --init $(awk '/path/ && !/llvm/ { print $3 }' .gitmodules)
 
-# Download and extract the Qt packages
-# Use caching to skip the download if we can
-for url in $QT_URLS; do
+# Git bash doesn't have rev, so here it is
+rev()
+{
+    echo "$1" | awk '{ for(i = length($0); i != 0; --i) { a = a substr($0, i, 1); } } END { print a }'
+}
+
+# Usage: download_and_verify url checksum algo file
+# Check to see if its already cached, and the checksum matches. If not, download it.
+# Tries up to 3 times
+download_and_verify()
+{
+    url="$1"
+    correctChecksum="$2"
+    algo="$3"
+    fileName="$4"
+
+    for _ in 1 2 3; do
+        [ -e "$CACHE_DIR/$fileName" ] || curl -L -o "$CACHE_DIR/$fileName" "$url"
+        fileChecksum=$("${algo}sum" "$CACHE_DIR/$fileName" | awk '{ print $1 }')
+        [ "$fileChecksum" = "$correctChecksum" ] && return 0
+        rm
+    done
+
+    return 1;
+}
+
+# Some dependencies install here
+[ -d "./lib" ] || mkdir "./lib"
+
+for url in $DEP_URLS; do
     # Get the filename from the URL. Breaks if urls have js args, so don't do that pls
     fileName="$(rev "$(rev "$url" | cut -d'/' -f1)")"
     [ -z "$fileName" ] && echo "Unable to parse url: $url" && exit 1
 
-    # Check to see if its already cached. If not, download it
-    if [ ! -e "$CACHE_DIR/$fileName" ]; then
-        curl -L -o "$CACHE_DIR/$fileName" "$url"
-    fi
+    # shellcheck disable=SC1003
+    case "$url" in
+    *qt*) checksum=$(curl -L "${url}.sha1"); algo="sha1"; outDir='C:\Qt\' ;;
+    *llvm*) checksum="$LLVM_SHA"; algo="sha256"; outDir="." ;;
+    *glslang*) checksum="$GLSLANG_SHA"; algo=sha256; outDir="./lib/Release - LLVM-x64" ;;
+    *Vulkan*)
+        # Vulkan setup needs to be run in batch environment
+        # Need to subshell this or else it doesn't wait
+        download_and_verify "$url" "$VULKAN_SDK_SHA" "sha256" "$fileName"
+        cp "$CACHE_DIR/$fileName" .
+        _=$(echo "$fileName /S" | cmd)
+        continue
+    ;;
+    *) echo "Unknown url resource: $url"; exit 1 ;;
+    esac
 
-    7z x -y "$CACHE_DIR/$fileName" -o'C:\Qt\'
+    download_and_verify "$url" "$checksum" "$algo" "$fileName"
+    7z x -y "$CACHE_DIR/$fileName" -aos -o"$outDir"
 done
-
-# Download other dependencies if they aren't cached
-[ -e "$CACHE_DIR/llvmlibs.7z" ] || curl -L -o "$CACHE_DIR/llvmlibs.7z" "$LLVMLIBS"
-[ -e "$CACHE_DIR/glslang.7z" ] || curl -L -o "$CACHE_DIR/glslang.7z" "$GLSLANG"
-[ -e "$CACHE_DIR/vulkansdk.exe" ] || curl -L -o "$CACHE_DIR/vulkansdk.exe" "$VULKAN_SDK_URL"
-
-vkChecksum=$(sha256sum "$CACHE_DIR/vulkansdk.exe" | awk '{ print $1 }')
-[ "$vkChecksum" = "$VULKAN_SDK_SHA" ] || (echo "Bad Vulkan SDK" && exit 1)
-
-# Extract the non-Qt dependencies from the cache into the proper build area
-[ -d "./lib" ] || mkdir "./lib"
-7z x -y "$CACHE_DIR"/llvmlibs.7z -aos -o"."
-7z x -y "$CACHE_DIR"/glslang.7z -aos -o"./lib/Release - LLVM-x64"
-
-# Vulkan setup needs to be run in batch environment
-# Need to subshell this or else it doesn't wait
-cp "$CACHE_DIR/vulkansdk.exe" .
-toss=$(echo "vulkansdk.exe /S" | cmd)
 
 # Gather explicit version number and number of commits
 COMM_TAG=$(awk '/version{.*}/ { printf("%d.%d.%d", $5, $6, $7) }' ./rpcs3/rpcs3_version.cpp)
